extract `AnimationWrapper`
