+++ src/wrapper/EvalWrapper.ts
@@ -6,6 +6,7 @@ import {
   type TTrace,
 } from './TraceUtil.ts';
 import { trim2microsecond } from '../api/time.ts';
+import type { TPanelMap } from '../api/settings.ts';
 
 export type TEvalHistory = {
   traceId: string;
@@ -105,8 +106,10 @@ export class EvalWrapper {
     // noop - it's impossible to restore native eval afterwards
   }
 
-  collectHistory() {
-    return Array.from(this.evalHistory.values());
+  collectHistory(panels: TPanelMap) {
+    return panels.eval.visible && panels.eval.wrap
+      ? Array.from(this.evalHistory.values())
+      : null;
   }
 
   cleanHistory() {
