+++ src/wrapper/Wrapper.ts
@@ -3,8 +3,6 @@ import {
   clearTimeout,
   setInterval,
   clearInterval,
-  requestAnimationFrame,
-  cancelAnimationFrame,
   TAG_EVAL_RETURN_SET_TIMEOUT,
   TAG_EVAL_RETURN_SET_INTERVAL,
   TAG_MISSFORTUNE,
@@ -12,7 +10,7 @@ import {
 import { TAG_EXCEPTION } from '../api/clone.ts';
 import {
   EWrapperCallstackType,
-  type TPanelVisibilityMap,
+  type TPanelMap,
   type TSettings,
 } from '../api/settings.ts';
 import { trim2microsecond, callingOnce } from '../api/time.ts';
@@ -29,6 +27,11 @@ import {
   type TCancelIdleCallbackHistory,
   type TRequestIdleCallbackHistory,
 } from './IdleWrapper.ts';
+import {
+  AnimationWrapper,
+  type TCancelAnimationFrameHistory,
+  type TRequestAnimationFrameHistory,
+} from './AnimationWrapper.ts';
 
 export enum ETimerType {
   TIMEOUT,
@@ -64,25 +67,6 @@ export type TClearTimerHistory = {
   handler: number | string;
   delay: number | undefined | string;
 };
-export type TRequestAnimationFrameHistory = {
-  traceId: string;
-  trace: TTrace[];
-  traceDomain: ETraceDomain;
-  calls: number;
-  handler: number | undefined | string;
-  selfTime: number | null;
-  online: number;
-  canceledCounter: number;
-  canceledByTraceIds: string[] | null;
-  fps: number;
-};
-export type TCancelAnimationFrameHistory = {
-  traceId: string;
-  trace: TTrace[];
-  traceDomain: ETraceDomain;
-  calls: number;
-  handler: number | undefined | string;
-};
 
 export type TWrapperMetrics = {
   onlineTimers: TOnlineTimerMetrics[] | null;
@@ -114,42 +98,30 @@ export class Wrapper {
   traceUtil = new TraceUtil();
   apiEval: EvalWrapper;
   // apiTimer: TimerWrapper;
-  // apiAnimation: AnimationWrapper;
+  apiAnimation: AnimationWrapper;
   apiIdle: IdleWrapper;
-
   onlineTimers: Map</*handler*/ number, TOnlineTimerMetrics> = new Map();
   setTimeoutHistory: Map</*traceId*/ string, TSetTimerHistory> = new Map();
   clearTimeoutHistory: Map</*traceId*/ string, TClearTimerHistory> = new Map();
   setIntervalHistory: Map</*traceId*/ string, TSetTimerHistory> = new Map();
   clearIntervalHistory: Map</*traceId*/ string, TClearTimerHistory> = new Map();
-  animationCallsMap = new Map</*traceId*/ string, /*calls*/ number>();
-  onlineAnimationFrameLookup: Map</*handler*/ number, /*traceId*/ string> =
-    new Map();
-  rafHistory: Map</*traceId*/ string, TRequestAnimationFrameHistory> =
-    new Map();
-  cafHistory: Map</*traceId*/ string, TCancelAnimationFrameHistory> = new Map();
   callCounter = {
-    activeTimers: 0,
     setTimeout: 0,
     clearTimeout: 0,
     setInterval: 0,
     clearInterval: 0,
-    requestAnimationFrame: 0,
-    cancelAnimationFrame: 0,
   };
   native = {
     setTimeout: setTimeout,
     clearTimeout: clearTimeout,
     setInterval: setInterval,
     clearInterval: clearInterval,
-    requestAnimationFrame: requestAnimationFrame,
-    cancelAnimationFrame: cancelAnimationFrame,
   };
 
   constructor() {
     this.apiEval = new EvalWrapper(this.traceUtil);
-    // this.apiTimer = new ApiTimer(this.traceUtil, this.apiEval);
-    // this.apiAnimationFrame = new ApiAnimationFrame(this.traceUtil);
+    // this.apiTimer = new TimerWrapper(this.traceUtil, this.apiEval);
+    this.apiAnimation = new AnimationWrapper(this.traceUtil);
     this.apiIdle = new IdleWrapper(this.traceUtil);
   }
 
@@ -296,118 +268,25 @@ export class Wrapper {
     }
   }
 
-  updateAnimationsFramerate() {
-    for (let [, rafRecord] of this.rafHistory) {
-      const prevCalls = this.animationCallsMap.get(rafRecord.traceId) || 0;
-      const fps = rafRecord.calls - prevCalls;
-
-      this.animationCallsMap.set(rafRecord.traceId, rafRecord.calls);
-      rafRecord.fps = fps;
-    }
-  }
-
-  updateRafHistory(handler: number, callstack: TCallstack) {
-    const existing = this.rafHistory.get(callstack.traceId);
-
-    if (existing) {
-      existing.calls++;
-      existing.handler = handler;
-      existing.online++;
-    } else {
-      this.rafHistory.set(callstack.traceId, {
-        traceId: callstack.traceId,
-        trace: callstack.trace,
-        traceDomain: this.traceUtil.getTraceDomain(callstack.trace[0]),
-        calls: 1,
-        handler,
-        online: 1,
-        canceledCounter: 0,
-        canceledByTraceIds: null,
-        selfTime: null,
-        fps: 1,
-      });
-    }
-
-    this.onlineAnimationFrameLookup.set(handler, callstack.traceId);
-  }
-
-  rafFired(handler: number, traceId: string, selfTime: number | null) {
-    const rafRecord = this.rafHistory.get(traceId);
-    if (!rafRecord) {
-      return;
-    }
-
-    rafRecord.selfTime = trim2microsecond(selfTime);
-
-    if (this.onlineAnimationFrameLookup.has(handler)) {
-      this.onlineAnimationFrameLookup.delete(handler);
-      rafRecord.online--;
-    }
-  }
-
-  updateCafHistory(handler: number | string, callstack: TCallstack) {
-    const existing = this.cafHistory.get(callstack.traceId);
-    const hasError = !validHandler(handler);
-
-    if (hasError) {
-      handler = TAG_EXCEPTION(handler);
-    }
-
-    if (existing) {
-      existing.calls++;
-      existing.handler = handler;
-    } else {
-      this.cafHistory.set(callstack.traceId, {
-        traceId: callstack.traceId,
-        trace: callstack.trace,
-        traceDomain: this.traceUtil.getTraceDomain(callstack.trace[0]),
-        calls: 1,
-        handler,
-      });
-    }
-
-    const rafTraceId = this.onlineAnimationFrameLookup.get(Number(handler));
-    const rafRecord = rafTraceId && this.rafHistory.get(rafTraceId);
-    if (rafRecord) {
-      this.onlineAnimationFrameLookup.delete(Number(handler));
-
-      rafRecord.online--;
-
-      if (rafRecord.canceledByTraceIds === null) {
-        rafRecord.canceledByTraceIds = [callstack.traceId];
-      } else if (!rafRecord.canceledByTraceIds.includes(callstack.traceId)) {
-        rafRecord.canceledByTraceIds.push(callstack.traceId);
-      }
-      rafRecord.canceledCounter++;
-    }
-  }
-
   cleanHistory() {
     this.apiEval.cleanHistory();
     this.apiIdle.cleanHistory();
-
+    this.apiAnimation.cleanHistory();
     this.setTimeoutHistory.clear();
     this.clearTimeoutHistory.clear();
     this.setIntervalHistory.clear();
     this.clearIntervalHistory.clear();
-    this.rafHistory.clear();
-    this.cafHistory.clear();
-    this.onlineAnimationFrameLookup.clear();
-    this.animationCallsMap.clear();
+
     this.callCounter.setTimeout =
       this.callCounter.clearTimeout =
       this.callCounter.setInterval =
       this.callCounter.clearInterval =
-      this.callCounter.requestAnimationFrame =
-      this.callCounter.cancelAnimationFrame =
         0;
   }
 
-  collectWrapperMetrics(panels: TPanelVisibilityMap): TWrapperMetrics {
-    this.callCounter.activeTimers = this.onlineTimers.size;
-
+  collectWrapperMetrics(panels: TPanelMap): TWrapperMetrics {
     return {
-      evalHistory: panels.eval.visible ? this.apiEval.collectHistory() : null,
+      evalHistory: this.apiEval.collectHistory(panels),
       onlineTimers: panels.activeTimers.visible
         ? Array.from(this.onlineTimers.values())
         : null,
@@ -423,38 +302,36 @@ export class Wrapper {
       clearIntervalHistory: panels.clearInterval.visible
         ? Array.from(this.clearIntervalHistory.values())
         : null,
-      rafHistory: panels.requestAnimationFrame.visible
-        ? Array.from(this.rafHistory.values())
-        : null,
-      cafHistory: panels.cancelAnimationFrame.visible
-        ? Array.from(this.cafHistory.values())
-        : null,
+      ...this.apiAnimation.collectHistory(panels),
       ...this.apiIdle.collectHistory(panels),
 
       callCounter: {
+        activeTimers: this.onlineTimers.size,
         eval: this.apiEval.callCounter,
-        requestIdleCallback: this.apiIdle.callCounter.requestIdleCallback,
-        cancelIdleCallback: this.apiIdle.callCounter.cancelIdleCallback,
+        ...this.apiAnimation.callCounter,
+        ...this.apiIdle.callCounter,
         ...this.callCounter,
       },
     };
   }
 
-  setup(panels: TPanelVisibilityMap, settings: TSettings) {
+  setup(panels: TPanelMap, settings: TSettings) {
     this.traceUtil.trace4Debug = settings.traceForDebug;
     this.traceUtil.trace4Bypass = settings.traceForBypass;
     this.setCallstackType(settings.wrapperCallstackType);
     this.wrapByPanels(panels);
   }
 
-  wrapByPanels = callingOnce((panels: TPanelVisibilityMap) => {
+  wrapByPanels = callingOnce((panels: TPanelMap) => {
     panels.eval.wrap && this.apiEval.wrap();
     panels.setTimeout.wrap && this.wrapSetTimeout();
     panels.clearTimeout.wrap && this.wrapClearTimeout();
     panels.setInterval.wrap && this.wrapSetInterval();
     panels.clearInterval.wrap && this.wrapClearInterval();
-    panels.requestAnimationFrame.wrap && this.wrapRequestAnimationFrame();
-    panels.cancelAnimationFrame.wrap && this.wrapCancelAnimationFrame();
+    panels.requestAnimationFrame.wrap &&
+      this.apiAnimation.wrapRequestAnimationFrame();
+    panels.cancelAnimationFrame.wrap &&
+      this.apiAnimation.wrapCancelAnimationFrame();
     panels.requestIdleCallback.wrap && this.apiIdle.wrapRequestIdleCallback();
     panels.cancelIdleCallback.wrap && this.apiIdle.wrapCancelIdleCallback();
   });
@@ -465,8 +342,8 @@ export class Wrapper {
     this.wrapClearTimeout();
     this.wrapSetInterval();
     this.wrapClearInterval();
-    this.wrapRequestAnimationFrame();
-    this.wrapCancelAnimationFrame();
+    this.apiAnimation.wrapRequestAnimationFrame();
+    this.apiAnimation.wrapCancelAnimationFrame();
     this.apiIdle.wrapRequestIdleCallback();
     this.apiIdle.wrapCancelIdleCallback();
   }
@@ -477,59 +354,10 @@ export class Wrapper {
     window.clearTimeout = this.native.clearTimeout;
     window.setInterval = this.native.setInterval;
     window.clearInterval = this.native.clearInterval;
-    window.requestAnimationFrame = this.native.requestAnimationFrame;
-    window.cancelAnimationFrame = this.native.cancelAnimationFrame;
+    this.apiAnimation.unwrapRequestAnimationFrame();
+    this.apiAnimation.unwrapCancelAnimationFrame();
     this.apiIdle.unwrapRequestIdleCallback();
     this.apiIdle.unwrapCancelIdleCallback();
-  }
-
-  wrapRequestAnimationFrame() {
-    window.requestAnimationFrame = function requestAnimationFrame(
-      this: Wrapper,
-      fn: FrameRequestCallback
-    ) {
-      const err = new Error(TraceUtil.SIGNATURE);
-      const callstack = this.traceUtil.createCallstack(err, fn);
-
-      this.callCounter.requestAnimationFrame++;
-      const handler = this.native.requestAnimationFrame((...args) => {
-        const start = performance.now();
-        let selfTime = null;
-
-        if (this.traceUtil.shouldPass(callstack.traceId)) {
-          if (this.traceUtil.shouldPause(callstack.traceId)) {
-            debugger;
-          }
-          fn(...args);
-          selfTime = performance.now() - start;
-        }
-
-        this.rafFired(handler, callstack.traceId, selfTime);
-      });
-      this.updateRafHistory(handler, callstack);
-
-      return handler;
-    }.bind(this);
-  }
-
-  wrapCancelAnimationFrame() {
-    window.cancelAnimationFrame = function cancelAnimationFrame(
-      this: Wrapper,
-      handler: number
-    ) {
-      const err = new Error(TraceUtil.SIGNATURE);
-      const callstack = this.traceUtil.createCallstack(err);
-
-      this.updateCafHistory(handler, callstack);
-      this.callCounter.cancelAnimationFrame++;
-
-      if (this.traceUtil.shouldPass(callstack.traceId)) {
-        if (this.traceUtil.shouldPause(callstack.traceId)) {
-          debugger;
-        }
-        this.native.cancelAnimationFrame(handler);
-      }
-    }.bind(this);
   }
 
   wrapSetTimeout() {
